Switch to cmake with support for Qt5/KF5
This is a work-in-progress, but most of the necessary pieces are in place. Here
is a brief summary of the changes:

- Detect the required Qt5/KF5 dependencies
- Let cmake take care of moc, uic, and qrc files
- Install files to standard locations

Note that we explicitly require Qt5/KF5 components, so there is no backwards
compatibility with earlier versions.

To build kfilebox using cmake (rather than qmake) please do the following:

$ mkdir build
$ cd build
$ cmake [options] ..
$ make