Rename locale *.po files and place in po/
It seems to be a common convention to place all po files into a po/ directory
and name them <lang>.po. Doing so greatly simplifies installation for me, and
hopefully for packagers.

One *extremely* important note: while the files seem to be correctly processed
into mo files, I was not able to get translations to work in my testing. For
example, setting LANG="de" translates none of the strings in kfilebox. Even the
previously generated mo files do not work for me.